+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Enhancements
 * Added `Text` operator support. (PR [#4](https://github.com/nirinchev/realm-dotnet-search/pull/4))
+* Added `Phrase` operator support. (PR [#4](https://github.com/nirinchev/realm-dotnet-search/pull/4))
+* Added some geojson types. (PR [#4](https://github.com/nirinchev/realm-dotnet-search/pull/4))
 
 ### Fixed
 * None